refactor(core): 1 util per file = no named exports = smaller bundle
