+++ packages/core/client.js
@@ -2,7 +2,9 @@ const config = require('./config')
 const Event = require('./event')
 const Breadcrumb = require('./breadcrumb')
 const Session = require('./session')
-const { map, includes, filter } = require('./lib/es-utils')
+const map = require('./lib/es-utils/map')
+const includes = require('./lib/es-utils/includes')
+const filter = require('./lib/es-utils/filter')
 const runCallbacks = require('./lib/callback-runner')
 const metadataDelegate = require('./lib/metadata-delegate')
 const runSyncCallbacks = require('./lib/sync-callback-runner')
