+++ packages/core/config.js
@@ -1,4 +1,8 @@
-const { filter, reduce, keys, isArray, includes } = require('./lib/es-utils')
+const filter = require('./lib/es-utils/filter')
+const reduce = require('./lib/es-utils/reduce')
+const keys = require('./lib/es-utils/keys')
+const isArray = require('./lib/es-utils/is-array')
+const includes = require('./lib/es-utils/includes')
 const intRange = require('./lib/validators/int-range')
 const stringWithLength = require('./lib/validators/string-with-length')
 const listOfFunctions = require('./lib/validators/list-of-functions')
