+++ packages/core/event.js
@@ -1,7 +1,9 @@
 const ErrorStackParser = require('./lib/error-stack-parser')
 const StackGenerator = require('stack-generator')
 const hasStack = require('./lib/has-stack')
-const { map, reduce, filter } = require('./lib/es-utils')
+const map = require('./lib/es-utils/map')
+const reduce = require('./lib/es-utils/reduce')
+const filter = require('./lib/es-utils/filter')
 const jsRuntime = require('./lib/js-runtime')
 const metadataDelegate = require('./lib/metadata-delegate')
 const isError = require('./lib/iserror')
