+++ packages/core/lib/es-utils/includes.js
@@ -1,0 +1,4 @@
+const reduce = require('./reduce')
+// Array#includes
+module.exports = (arr, x) =>
+  reduce(arr, (accum, item, i, arr) => accum === true || item === x, false)
