+++ packages/core/lib/es-utils/is-array.js
@@ -1,0 +1,2 @@
+// Array#isArray
+module.exports = obj => Object.prototype.toString.call(obj) === '[object Array]'
