+++ packages/expo/src/config.js
@@ -1,7 +1,6 @@
 /* global __DEV__ */
 
 const { schema } = require('@bugsnag/core/config')
-const { reduce } = require('@bugsnag/core/lib/es-utils')
 const Constants = require('expo-constants').default
 
 // If the developer property is not present in the manifest, it means the app is
@@ -30,7 +29,7 @@ module.exports = {
 }
 
 const getPrefixedConsole = () => {
-  return reduce(['debug', 'info', 'warn', 'error'], (accum, method) => {
+  return ['debug', 'info', 'warn', 'error'].reduce((accum, method) => {
     // console.error causes standalone expo apps to reload on android
     // so don't do any logging that level – use console.warn instead
     const consoleMethod = (IS_PRODUCTION && method === 'error') ? console.warn : console[method]
