+++ packages/plugin-navigation-breadcrumbs/navigation-breadcrumbs.js
@@ -1,4 +1,4 @@
-const { includes } = require('@bugsnag/core/lib/es-utils')
+const includes = require('@bugsnag/core/lib/es-utils/includes')
 
 /*
  * Leaves breadcrumbs when navigation methods are called or events are emitted
